Update preprocess.py
Dear X, I have wrote the function `rebin(sightline, v)` to rebin the data to a constant dlambda/lambda. I referenced the code in linetools and made a little change. I think maybe it's better to add this function to Sightline class, but that doesn't matter.   Hope for your advice!